+++ app/src/main/java/com/etherealhazel/stakediv/container/ContainerRepository.java
@@ -1,8 +1,6 @@
-package com.etherealhazel.stakediv.repo;
+package com.etherealhazel.stakediv.container;
 
 import java.util.UUID;
-
-import com.etherealhazel.stakediv.model.Container;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
