+++ app/src/main/java/com/etherealhazel/stakediv/container/EmbeddedContainerDto.java
@@ -1,4 +1,4 @@
-package com.etherealhazel.stakediv.dto;
+package com.etherealhazel.stakediv.container;
 
 import java.util.UUID;
