+++ app/src/main/java/com/etherealhazel/stakediv/user/EmbeddedUserSerializer.java
@@ -1,12 +1,10 @@
 //https://docs.spring.io/spring-boot/docs/2.3.8.RELEASE/reference/html/spring-boot-features.html#boot-features-json-components
-package com.etherealhazel.stakediv.serializer;
+package com.etherealhazel.stakediv.user;
 
 import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.etherealhazel.stakediv.dto.EmbeddedUserDto;
-import com.etherealhazel.stakediv.model.AppUser;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.SerializerProvider;
