+++ app/src/main/java/com/etherealhazel/stakediv/user/UserDto.java
@@ -1,4 +1,4 @@
-package com.etherealhazel.stakediv.dto;
+package com.etherealhazel.stakediv.user;
 
 public class UserDto {
